+++ package.json
@@ -14,7 +14,8 @@
     "@headlessui/tailwindcss": "^0.2.0",
     "@heroicons/react": "^2.1.3",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.23.0"
   },
   "devDependencies": {
     "@types/react": "^18.2.66",
