+++ src/main.jsx
@@ -4,53 +4,7 @@ import App from './App.jsx'
 import './index.css'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
-  <main>
-  <App />
-  <div className="w-full bg-[#020C0E]">
-      <div className="max-w-lg space-y-4 mx-auto py-24 text-neutral-300">
-        <p>
-          Lorem ipsum dolor sit amet consectetur, adipisicing elit. Ipsam quae
-          earum nobis quasi repellat. Amet facere nulla dolorum accusantium
-          sit dolores odio excepturi facilis laboriosam officiis dolorem,
-          nobis reprehenderit molestiae.
-        </p>
-        <p>
-          Lorem ipsum dolor sit amet consectetur, adipisicing elit. Ipsam quae
-          earum nobis quasi repellat. Amet facere nulla dolorum accusantium
-          sit dolores odio excepturi facilis laboriosam officiis dolorem,
-          nobis reprehenderit molestiae.
-        </p>
-        <p>
-          Lorem ipsum dolor sit amet consectetur, adipisicing elit. Ipsam quae
-          earum nobis quasi repellat. Amet facere nulla dolorum accusantium
-          sit dolores odio excepturi facilis laboriosam officiis dolorem,
-          nobis reprehenderit molestiae.
-        </p>
-        <p>
-          Lorem ipsum dolor sit amet consectetur, adipisicing elit. Ipsam quae
-          earum nobis quasi repellat. Amet facere nulla dolorum accusantium
-          sit dolores odio excepturi facilis laboriosam officiis dolorem,
-          nobis reprehenderit molestiae.
-        </p>
-        <p>
-          Lorem ipsum dolor sit amet consectetur, adipisicing elit. Ipsam quae
-          earum nobis quasi repellat. Amet facere nulla dolorum accusantium
-          sit dolores odio excepturi facilis laboriosam officiis dolorem,
-          nobis reprehenderit molestiae.
-        </p>
-        <p>
-          Lorem ipsum dolor sit amet consectetur, adipisicing elit. Ipsam quae
-          earum nobis quasi repellat. Amet facere nulla dolorum accusantium
-          sit dolores odio excepturi facilis laboriosam officiis dolorem,
-          nobis reprehenderit molestiae.
-        </p>
-        <p>
-          Lorem ipsum dolor sit amet consectetur, adipisicing elit. Ipsam quae
-          earum nobis quasi repellat. Amet facere nulla dolorum accusantium
-          sit dolores odio excepturi facilis laboriosam officiis dolorem,
-          nobis reprehenderit molestiae.
-        </p>
-      </div>
-    </div>
-  </main>
+  <>
+    <App />
+  </>
 )
