+++ src/pages/Browse/index.jsx
@@ -1,0 +1,15 @@
+function Browse() {
+    
+    // HTML
+    return (
+        <>
+            <h1>
+                BROWSE
+            </h1>
+        </>
+
+    );
+}
+
+// Exports the Browse page
+export default Browse;
