+++ src/pages/Home/index.jsx
@@ -1,0 +1,15 @@
+function Home() {
+    
+    // HTML
+    return (
+        <>
+            <h1>
+                HOME
+            </h1>
+        </>
+
+    );
+}
+
+// Exports the Home page
+export default Home;
