+++ src/pages/Profile/index.jsx
@@ -1,0 +1,15 @@
+function Profile() {
+    
+    // HTML
+    return (
+        <>
+            <h1>
+                PROFILE
+            </h1>
+        </>
+
+    );
+}
+
+// Exports the Profile page
+export default Profile;
