+++ src/pages/Search/index.jsx
@@ -1,0 +1,15 @@
+function Search() {
+    
+    // HTML
+    return (
+        <>
+            <h1>
+                SEARCH
+            </h1>
+        </>
+
+    );
+}
+
+// Exports the Search page
+export default Search;
